Tier 3a: EIPの32bit化 — 最後の16bit配線を外す (#28)
* nasmのリストファイルを追い出す (デバッグ中の生成物がPR #26に紛れた)

* EIPを32bitにする — 最後の16bit配線を外す

ip: u16 はコアに残った最後の16bit配線だった。64KBを超えるコード
(Linuxカーネルは1MB超) はこれが外れないと歩けない。

## 折り返しの持ち主を1箇所にする

widen して終わりではなく、「16bitコードでは64Kで折り返す」という8086の
挙動を**どこが知っているか**を決めるのが本体だった。散らばらせず、
Cpu::ip_mask() / set_ip() / advance_ip() の3つだけが幅を知る。
幅はモードとCSのDビットから導く — regs を最初から u32 で持っていたのと
同じ判断を IP にも適用した形。

- フェッチ (最熱経路) は advance_ip(1)。ANDが1個増えるだけ
- 分岐族 (jmp/call/ret/jcc/loop/grp5) は着地を set_ip に統一。
  ついでに**オペランドサイズ対応** (rel32 / pop32) まで済ませた —
  型を直すだけなら二度手間になる場所
- far call/retf (9A/CB/CA) はリアルモード16bit形のまま。保護モードの
  far call は特権チェックと絡むので、リングを作るときに一緒にやる

## 道連れ

- 割り込み統計 (int_first/int_recent/first_fault) のIPを u32 に
- debug (Step.ip / Stop::*.at / before_exec) を u32 に
- スナップショット v4
- boot.rs の診断にあった sel<<4 直書きの残党2箇所も cpu.lin() に統一

## 検証

- 新テスト2件: **EIPが64Kを越えて歩ける** (0x2_0000 のコードを実行。
  これまでの全テストは偶然64K以下で走っていた) / **リアルモードのIPは
  64Kで折り返す** (8086の環)
- 103テスト全通過、clippy 0
- **ELKSがまたビット同一 (45,669,277命令)** — フェッチ経路を触っても
  リアルモードの挙動はゼロ変化。FreeDOS も dir まで無傷

* cosim を EIP 32bit化に追随させる

CPU照合 (CI) だけが捕まえたビルド漏れ。cosim は default-members の外に
居るので、手元の cargo build --all-targets も clippy --workspace
--exclude rustx86-cosim も、全部この crate を素通りしていた。
Unicorn のCビルドが重いから除外している以上、ここを見張れるのは
paths で発火する CPU照合 ワークフローだけ — 設計どおりの網に掛かった。

State::ip は u16 のまま、渡す側を切り詰める。MODE_16 の突き合わせでは
IPは16bitが正しい。MODE_32 対応のときに器ごと広げる。

手元でも cargo test -p rustx86-cosim を実際に回した: 15件全通過
(DAAの総当たり含む、112秒)。

+++ core/src/debug.rs
@@ -51,12 +51,12 @@ pub enum Stop {
         addr: u32,
         old: u8,
         new: u8,
-        at: (u16, u16),
+        at: (u16, u32),
     },
     /// I/O書き込み
-    WriteIo { port: u16, val: u8, at: (u16, u16) },
+    WriteIo { port: u16, val: u8, at: (u16, u32) },
     /// I/O読み出し
-    ReadIo { port: u16, val: u8, at: (u16, u16) },
+    ReadIo { port: u16, val: u8, at: (u16, u32) },
     /// 指定した命令数に達した
     Count(u64),
 }
@@ -67,7 +67,7 @@ pub enum Stop {
 pub struct Step {
     pub instr: u64,
     pub cs: u16,
-    pub ip: u16,
+    pub ip: u32,
     /// 命令の先頭5バイト。逆アセンブラはまだ無いので生のまま持つ
     pub bytes: [u8; 5],
 }
@@ -109,7 +109,7 @@ pub struct Debug {
     /// 止まった理由。**呼び出し側が取り去るまで残る**
     pub stop: Option<Stop>,
     /// いま実行中の命令の先頭。フックが「誰が書いたか」を答えるための元
-    pub at: (u16, u16),
+    pub at: (u16, u32),
     /// 実行の足跡。`trace_cap` が0なら何も残さない
     pub trace: VecDeque<Step>,
     pub trace_cap: usize,
@@ -218,7 +218,7 @@ impl Debug {
     /// **実行前**に判定するので、止まった状態でその命令を見られる。
     ///
     /// 止まっている (HLT) 間は呼ばれない。呼ぶと同じ番地で永久に止まるため
-    pub(crate) fn before_exec(&mut self, lin: u32, cs: u16, ip: u16) -> bool {
+    pub(crate) fn before_exec(&mut self, lin: u32, cs: u16, ip: u32) -> bool {
         self.at = (cs, ip);
         if self.code.is_empty() {
             return false;
